refactor: get offers from store in all components
