+++ src/components/app/app.tsx
@@ -1,15 +1,17 @@
 import { Route, BrowserRouter, Routes } from 'react-router-dom';
-import { AppRoute} from '../../consts/consts';
+import { AppRoute, AuthorizationStatus} from '../../consts/consts';
 import MainPage from '../../pages/main-page/main-page';
 import LoginPage from '../../pages/login-page/login-page';
 import OfferPage from '../../pages/offer-page/offer-page';
 import FavoritesPage from '../../pages/favorites-page/favorites';
 import NotFoundPage from '../../pages/not-found-page/not-found-page';
-import PrivateRoute from '../private-route/pribate-route';
+import PrivateRoute from '../private-route/private-route';
 import { HelmetProvider } from 'react-helmet-async';
 
-
 function App(): JSX.Element {
+
+  const authorizationStatus = AuthorizationStatus.Auth;
+
   return(
     <HelmetProvider>
       <BrowserRouter>
@@ -24,13 +26,13 @@ function App(): JSX.Element {
           />
           <Route
             path={AppRoute.Room}
-            element={<OfferPage offers={offers} authorizationStatus={authorizationStatus} />}
+            element={<OfferPage authorizationStatus={authorizationStatus} />}
           />
           <Route
             path={AppRoute.Favorites}
             element={
               <PrivateRoute authorizationStatus={authorizationStatus}>
-                <FavoritesPage offers={offers}/>
+                <FavoritesPage />
               </PrivateRoute>
             }
           />
