+++ src/store/selectors/offers.ts
@@ -1,6 +1,6 @@
 import type { RootState } from '../../types/store';
 
-const selectOffers = (state: RootState) => state.offers;
-const selectCity = (state: RootState) => state.city;
+const selectOffers = (state: RootState) => state.offers.offers;
+const selectCity = (state: RootState) => state.offers.city;
 
 export {selectCity, selectOffers};
